Refactor to remove overloaded `target` terminology
- Renamed `repo_target.rs` to `repo_resolution.rs` and updated module exports.
- Updated `gh` labels commands to use `repo_resolution` imports.
- Renamed `target` local variable to `value_str` in `backup/system.rs` numeric formatter.
- Renamed `target` local variable to `dest_dir` in `deploy_configs.rs`.
- Renamed `BackupArgs` field `target` to `component`, maintaining CLI compatibility with `#[arg(name = "target")]`.
