+++ crates/mev-internal/src/app/commands/gh/labels_reset.rs
@@ -4,7 +4,7 @@ use clap::Args;
 
 use crate::adapters::{gh, git};
 use crate::domain::DomainError;
-use crate::domain::repo_target;
+use crate::domain::repo_resolution;
 
 #[derive(Args)]
 pub struct LabelsResetArgs {
@@ -16,7 +16,7 @@ pub struct LabelsResetArgs {
 pub fn run(args: LabelsResetArgs) -> Result<(), DomainError> {
     let git_adapter = git::GitAdapter::default();
     let origin_url = args.repo.is_none().then(|| git_adapter.current_origin_url()).transpose()?;
-    let repo = repo_target::resolve_repo_ref(args.repo.as_deref(), origin_url.as_deref())?;
+    let repo = repo_resolution::resolve_repo_ref(args.repo.as_deref(), origin_url.as_deref())?;
 
     let gh_adapter = gh::GhAdapter::default();
     let names = gh_adapter.list_label_names(&repo)?;
