+++ crates/mev-internal/src/domain/mod.rs
@@ -2,7 +2,7 @@
 
 pub mod error;
 pub mod label_catalog;
-pub mod repo_target;
+pub mod repo_resolution;
 pub mod repository_ref;
 pub mod submodule_path;
